requirements.txt
django
djangorestframework
qrcode
pillow
django-filter
markdown
django-cors-middleware